+++ apps/meet/app/(app)/layout.tsx
@@ -84,6 +84,7 @@ export default async function MeetAppLayout({
       </div>
       <div className="flex-1 flex flex-col min-w-0">
         <Topbar
+          todoEnabled={me.todo_enabled !== false}
           email={email}
           fullName={fullName}
           prefs={prefs}